+++ pyobjc-framework-WebKit/PyObjCTest/test_wkwebsitedatarecord.py
@@ -15,5 +15,11 @@ class TestWKWebsiteDataRecord (TestCase):
         self.assertIsInstance(WKWebsiteDataTypeWebSQLDatabases, unicode)
         self.assertIsInstance(WKWebsiteDataTypeIndexedDBDatabases, unicode)
 
+    @onlyOn64Bit
+    @min_os_level('10.13.4')
+    def testConstants10_13_4(self):
+        self.assertIsInstance(WKWebsiteDataTypeFetchCache, unicode)
+        self.assertIsInstance(WKWebsiteDataTypeServiceWorkerRegistrations, unicode)
+
 if __name__ == "__main__":
     main()